FS-5062 - Transferring tests of fund module from general test_db into separate test module
